+++ common/rfb/CMsgWriter.h
@@ -54,7 +54,7 @@ namespace rfb {
     void writeFence(uint32_t flags, unsigned len, const uint8_t data[]);
 
     void writeKeyEvent(uint32_t keysym, uint32_t keycode, bool down);
-    void writePointerEvent(const Point& pos, int buttonMask);
+    void writePointerEvent(const Point& pos, uint8_t buttonMask);
 
     void writeClientCutText(const char* str);
