+++ unix/x0vncserver/XDesktop.h
@@ -60,7 +60,7 @@ public:
   bool isRunning();
   void queryConnection(network::Socket* sock,
                        const char* userName) override;
-  void pointerEvent(const rfb::Point& pos, int buttonMask) override;
+  void pointerEvent(const rfb::Point& pos, uint8_t buttonMask) override;
   void keyEvent(uint32_t keysym, uint32_t xtcode, bool down) override;
   unsigned int setScreenLayout(int fb_width, int fb_height,
                                const rfb::ScreenSet& layout) override;
@@ -79,7 +79,7 @@ protected:
   rfb::VNCServer* server;
   QueryConnectDialog* queryConnectDialog;
   network::Socket* queryConnectSock;
-  int oldButtonMask;
+  uint8_t oldButtonMask;
   bool haveXtest;
   bool haveDamage;
   int maxButtons;
